Upgrade to isort v5 (fixes #8)
isort>=5 does not support the --recursive option (it works recursively
anyway), so it had to be removed. In addition, apparently it searches
from the current directory upwards for .isort.cfg, and not from the
checked file upwards, so we need to change directory in order to check
different working directories.