Revert "Revising Vagrant guide"
This reverts commit b35841bb78c2024c12abe153b398d5fbcbd6f681, which had
several problems. First of all, it isn't revising Vagrant guide at all;
it just changes the Vagrant file, attempting to remove the step of
cloning the Enhydris repositories. However, it makes assumptions about
the way the user wants to connect to github or about the existence of
github authorized ssh keys. It generally didn't work on Windows.