Add support for setting up a MacOS environment (fixes #18)
Update the `Dockerfile` and the `run.sh` script to work
in a Mac OS environment without breaking support to the
existing supported operating systems (i.e. Linux & Windows).

The README.md is also updated to address what should be done
when running a migration fails due to existing tables that is
imported by the `dbimport` command.